Abhi 12-02-24

+++ app/src/main/java/codeflies/com/pulse/Models/Leaves/LeavesDetails.kt
@@ -56,7 +56,7 @@ data class Leave(
     val employeeCode: String? = null,
 
     @field:SerializedName("leave_days")
-    val leaveDays: Int? = null,
+    val leaveDays: Double? = null,
 
     @field:SerializedName("approved_by")
     val approvedBy: Any? = null,
